+++ tests/files/with-resolved-types-dependencies/foo.js
@@ -1,0 +1,3 @@
+import qs from 'qs';
+
+export default qs;